Refactor Helpfulness & Report button into separate components
