+++ client/src/components/ratings-and-reviews/HelpfulnessDisplay.jsx
@@ -1,0 +1,27 @@
+import axios from 'axios';
+import React from 'react';
+
+const { useState, useRef } = React;
+
+export default function HelpfulnessDisplay({ id, helpfulness }) {
+  const [currentHelpfulness, setHelpfulness] = useState(helpfulness);
+  const clickedHelpful = useRef(false);
+  const handleHelpful = () => {
+    if (!clickedHelpful.current) {
+      axios.put(`/reviews/${id}/helpful`);
+      setHelpfulness(currentHelpfulness + 1);
+      clickedHelpful.current = true;
+    }
+  };
+  return (
+    <span className="helpfulness">
+      Helpful?
+      <button type="button" className="review-helpful" onClick={handleHelpful}>Yes</button>
+      <span className="review-helpfulness-display">
+        (
+        {currentHelpfulness}
+        )
+      </span>
+    </span>
+  );
+}
