+++ client/src/components/ratings-and-reviews/ReportButton.jsx
@@ -1,0 +1,19 @@
+import axios from 'axios';
+import React from 'react';
+
+const { useState } = React;
+
+export default function ReportButton({ id }) {
+  const [reported, setReported] = useState(false);
+  const handleReport = () => {
+    if (!reported) {
+      axios.put(`/reviews/${id}/report`);
+      setReported(true);
+    }
+  };
+  return (
+    <button type="button" className="report-review" onClick={handleReport}>
+      {reported ? 'Reported' : 'Report'}
+    </button>
+  );
+}
