+++ client/src/components/ratings-and-reviews/ReviewTile.jsx
@@ -1,15 +1,13 @@
-import axios from 'axios';
 import React from 'react';
 import StarRatings from '../shared/StarRatings';
+import HelpfulnessDisplay from './HelpfulnessDisplay';
+import ReportButton from './ReportButton';
 
-const { useState, useEffect, useRef } = React;
+const { useState, useEffect } = React;
 
 export default function ReviewTile({ review }) {
   const [body, setBody] = useState('');
-  const [helpfulness, setHelpfulness] = useState(review.helpfulness);
-  const [reported, setReported] = useState(false);
   const [showMore, setShowMore] = useState(false);
-  const clickedHelpful = useRef(false);
   useEffect(() => {
     if (review.body.length > 250) {
       setBody(`${review.body.slice(0, 251)}...`);
@@ -18,19 +16,6 @@ export default function ReviewTile({ review }) {
       setBody(review.body);
     }
   }, []);
-  const handleHelpful = () => {
-    if (!clickedHelpful.current) {
-      axios.put(`/reviews/${review.review_id}/helpful`);
-      setHelpfulness(helpfulness + 1);
-      clickedHelpful.current = true;
-    }
-  };
-  const handleReport = () => {
-    if (!reported) {
-      axios.put(`/reviews/${review.review_id}/report`);
-      setReported(true);
-    }
-  };
   const handleShowMore = () => {
     setShowMore(false);
     setBody(review.body);
@@ -62,38 +47,24 @@ export default function ReviewTile({ review }) {
         {summaryRemainder}
         {body}
       </div>
-      {
-        showMore
-          ? <button type="button" className="show-more" onClick={handleShowMore}>Show more</button>
-          : null
-      }
-      {
-      review.recommend
+      {showMore
+        ? <button type="button" className="show-more" onClick={handleShowMore}>Show more</button>
+        : null}
+      {review.recommend
         ? <div className="recommend-label">&#x2713; I recommend this product</div>
-        : null
-      }
-      {
-        review.response
-          ? (
-            <div className="review-response">
-              <div className="review-response-header">Response from seller:</div>
-              {review.response}
-            </div>
-          )
-          : null
-      }
+        : null}
+      {review.response
+        ? (
+          <div className="review-response">
+            <div className="review-response-header">Response from seller:</div>
+            {review.response}
+          </div>
+        )
+        : null}
       <div className="review-bottom-row">
-        Helpful?
-        <button type="button" className="review-helpful" onClick={handleHelpful}>Yes</button>
-        <span className="review-helpfulness-display">
-          (
-          {helpfulness}
-          )
-        </span>
+        <HelpfulnessDisplay id={review.review_id} helpfulness={review.helpfulness} />
         |
-        <button type="button" className="report-review" onClick={handleReport}>
-          {reported ? 'Reported' : 'Report'}
-        </button>
+        <ReportButton id={review.review_id} />
       </div>
     </div>
   );
